defiI-PolyAI/backend/requirements.txt
fastapi
uvicorn
pillow
sqlalchemy
transformers
torch
scikit-learn
joblib
python-multipart
